+++ ibrdtn/daemon/src/storage/DataStorage.cpp
@@ -90,7 +90,7 @@ namespace dtn
 		std::istream& DataStorage::istream::operator*()
 		{ return *_stream; }
 
-		DataStorage::DataStorage(Callback &callback, const ibrcommon::File &path, size_t write_buffer, bool initialize)
+		DataStorage::DataStorage(Callback &callback, const ibrcommon::File &path, unsigned int write_buffer, bool initialize)
 		 : _callback(callback), _path(path), _tasks(), _store_sem(write_buffer), _store_limited(write_buffer > 0), _faulty(false)
 		// limit the number of bundles in the write buffer
 		{
